Fixed some strict compiler warnings
Ticket: None
Changelog: None

+++ lib/sha1.c
@@ -146,7 +146,7 @@ int SHA1Result(SHA1Context *context, uint8_t Message_Digest[SHA1HashSize]) {
 
   for (i = 0; i < SHA1HashSize; ++i) {
     Message_Digest[i] =
-        context->Intermediate_Hash[i >> 2] >> 8 * (3 - (i & 0x03));
+        (uint8_t)(context->Intermediate_Hash[i >> 2] >> 8 * (3 - (i & 0x03)));
   }
 
   return shaSuccess;
@@ -358,14 +358,14 @@ void SHA1PadMessage(SHA1Context *context) {
   /*
    *  Store the message length as the last 8 octets
    */
-  context->Message_Block[56] = context->Length_High >> 24;
-  context->Message_Block[57] = context->Length_High >> 16;
-  context->Message_Block[58] = context->Length_High >> 8;
-  context->Message_Block[59] = context->Length_High;
-  context->Message_Block[60] = context->Length_Low >> 24;
-  context->Message_Block[61] = context->Length_Low >> 16;
-  context->Message_Block[62] = context->Length_Low >> 8;
-  context->Message_Block[63] = context->Length_Low;
+  context->Message_Block[56] = (uint8_t)(context->Length_High >> 24);
+  context->Message_Block[57] = (uint8_t)(context->Length_High >> 16);
+  context->Message_Block[58] = (uint8_t)(context->Length_High >> 8);
+  context->Message_Block[59] = (uint8_t)(context->Length_High);
+  context->Message_Block[60] = (uint8_t)(context->Length_Low >> 24);
+  context->Message_Block[61] = (uint8_t)(context->Length_Low >> 16);
+  context->Message_Block[62] = (uint8_t)(context->Length_Low >> 8);
+  context->Message_Block[63] = (uint8_t)(context->Length_Low);
 
   SHA1ProcessMessageBlock(context);
 }
